Fix conflict between usermove and go
According to the XBoard spec, an engine receiving "usermove" when not in
force mode should begin thinking immediately. Therefore, sending
"usermove" followed by "go" results in the engine picking two moves (one
for each side). So, the simplest change is to put the engine in force
mode before sending the usermove or setboard commands.