Fix problems with python-chess draw claims
In the python-chess library used by lichess-bot,

    board.game_is_over(claim_draw=True)

returns True not only according to the 50-move rule or the 3-fold rule,
but also if there is a legal move on the current that would activate
these rules. Although this is proper according to FIDE rules, lichess
does not have the facility for players (whether bot or human) to make
these pre-move claims. Lichess will not recognize a rule-based draw
until the actual move is made.

In this commit, instead of using the local board to determine if the
game is over, the engine is told to keep playing until lichess says the
game is over. This check is done by the new function `is_game_over()`
that checks whether the `status` field in the JSON message is set to
`started`.