Upgrade python chess library to latest version (1.4.0) (#260)
* Upgrade pychess library to latest version (1.3.3)

This commit updates lichess-bot to use the latest version of the python
chess library. The main result is a large decrease in the amount of code
in engine_wrapper.py due to so many functions being incorporated into
the chess library.

One annoying difference is that the new library rejects the 'usermove'
feature from Xboard engines, which is rare among chess engine
interfaces.

I hope this can be tested by a variety of engines to make sure it is on
par with the current lichess-bot. This would close #240.

* Remove managed options from engine configs

Some options are automatically managed by the chess engine interface, so
these need to be removed from the engine configurations before creating
the engine.

* Remove parentheses from non-callable members

In the latest version of the python chess library, the result returned
from book lookups has the Move result as a member, not a method, so it
is not callable.

* Remove option that is managed by chess library

This option is also redundant with uci_ponder.

* Use actual latest version of python-chess

* Fix UCI movetime usage

In config file, movetime is in milliseconds, but the library Limit class
expects seconds.

* Handle omitted go_commands

Line 84 prevents self.go_commands from being None

The movetime lines in search_with_ponder() check if there is a movetime
go_command and, if so, convert it from milliseconds to seconds.

* Cosmetic changes

- Delete trailing whitespace
- Standardize spacing around operators
- Delete unneeded parentheses

* Consolidate repeated code

* Fix handling of endgame table paths for Xboard

* Consolidate repeated stats code

* Consolidate repeated stderr logic

* Consolidate repeated move search code

The interface differences between UCI and Xboad protocols have been much
reduced in the latest python-chess library, allowing for much more code
to be shared here.