Fix Windows logging issues (#294)
* Start of multiprocessing logging solution

On Windows, logging everything does not work due to differences
in how Linux and Windows create new processes. This solution is
adapted from

https://docs.python.org/3/howto/logging-cookbook.html#logging-to-a-single-file-from-multiple-processes

Now, all that needs to happen is for Ctrl-C to be handled more
gracefully in Windows. This needs testing on Linux as well.

* Stop double-logging on Linux

* Better spacing for move numbers

* Gracefully exit from Ctrl-C in Windows

* Fix duplicate logging on some Windows machines

Some Windows machines end up with duplicate QueueHandlers
in the handler lists of game_start() handlers. This commit clears
all handlers before adding the QueueHandler.