Handle unexpected responses from lichess.org without a `type` field (#355)
* Handle unexpected responses from lichess.org

There are several open bug reports that have the error message

    Traceback (most recent call last):
      File "lichess-bot.py", line 131, in start
        if event["type"] == "terminated":
    KeyError: 'type'

These include #188, #197, #201, #224, #245, and most recently #311.

This commit checks that the incoming message has a `type` field
and, if it does not, prints an error message including the full
response from lichess. If the error is "Missing scope", the printed
message will include the warning that the API access token does
not have the permission to play as a bot.

I considered terminating lichess-bot when encountering this error,
but some of the bug reports mentioned that this error was
intermittent, so lichess-bot will remain connected to lichess.org.

* Separate error handling from normal code

Also, continue on all invalid events.

* Change error() to warning()

Seems more appropriate since lichess-bot might be able to
continue despite the unexpected response.