Improvements for PGN writing (#457)
* Get initial PGN record from lichess

Simplify PGN writing (especially the headers) by getting the data from lichess
directly. Lichess also has more information for some headers, like ELO and
openings

* Don't use title and rating in PGN file name

* More accurate comments

* Simplify PGN creation by downloading from lichess

The PGN record from lichess has more extensive header information, so use
that instead of manually parsing the last game JSON.

* Remove unnecessary header edit

* Comment change

* Adapt tests to new PGN function

* Add limitations to commentary-move alignment

Don't allow partial matches.

In case a set of moves happens more than once, favor the last matches.

* Simplify aligning the comments with the move list

* Handle when downloaded moves are the whole game

If a game is played without disconnection, all of the moves will be present
in the downloaded game. If the game has a disconnection and then is
completed on the next connection, the reloaded game record will not have
all moves. This commit handles both scenarios.

* Add clock data from lichess

* Consolidate repeated code

* Iterate over moves in one loop

* Inline unneeded local function

* Inline unneeded variable

* Move comment

* Simplify comment index calculation

Mostly for removing the arbitrary dependence on the board.

* Update game headers instead of replacing them

* Add reason for game ending as comment to last move

* Get lichess annotations

Put opening and endgame annotations from lichess into the PGN file.

* Revert "Add reason for game ending as comment to last move"

This reverts commit 4f83c8148ac26553ac2016d84da62f8acce549ea.

* Add new comments from lichess

* Remove newlines from between joined comments

Lichess sends moves with multiple comments, these comments are
separately delimited: {First} {Second}. The default PGN parser in
python-chess joins the comments with a newline. This commit creates
a slightly modified parser that joins these comments with a space.

There is a different change that could be done with a regex one-liner, but
that relies on both comments being contained in the same line.

* Simpler comment joiner

Use an f-string and just use a space to join instead of a slash.

* Fix error when db move is chosen after engine move

If a database or book move is chosen in the middle of a game and there are
engine-chosen moves before that move, then the engine's comments
will no longer be aligned with the moves on the board. This commit
preserves that alignment by inserting None as the comment to non-engine
moves after the first engine move.

The code was further cleaned up by creating new EngineWrapper methods.