Refactor for code readability (#506)
* Use f-strings instead of % syntax

* Delete trailing whitespace

* Split up dict definition into multiple lines

* Split up literal list into multiple lines

* Use triple-quoted string to split up literal

* Split function argument list into multiple lines

* Replace eval with converting object to bytes

This way, one gets the full benefit of syntax- and type-checking and
automatic conversion to strings. The .replace() call is required because
double-quotes are expected by the JSON consumer while the str()
converter uses single-quotes inside the returned string.

* Use triple-quotes to split long string

* Shorten lines with variables

* Split long function argument lists

* Replace list of strings with triple-quoted string

* No need to make a copy of the stockfish binary

It's fine for the same executable to be run twice at the same time.

* Make stockfish_path a global variable

Reduces the number of string computations and function argument count.

* Improve readability of offer_draw_or_resign

Use variables to cut down on code repetition and line lengths.

* Shorten line by splitting function arguments

* Use separate variable definition to shorten line

* Simplify trivially formatted string

* Simplify name/title/rating handling

Minimum lichess rating is not zero, so simpler checks work fine.

* Simpler Xboard EGT handling

* Combine lines for single variable

* Reduce repetition in EngineWrapper.get_stats()

* Split method arguments across multiple lines

These lines include some complex in-lined expressions, which can make
it hard to tell where one argument ends and the next begins. Other
methods with single variable arguments are left as is.

* Replace % syntax with f-string

* Split function arguments to different lines

* Delete trailing whitespace

* Use variables instead of dictionary access

Also, consolidate logic of correspondence disconnect time.

* Replace dictionary expressions with variables

* Shorten if-statement with variables

* Reduce code repetition in get_chessdb_move()

* Separate in-line dictionary access to variables

* Change order of branches to remove repetition

* Move in-line dictionary access to separate lines

* Shorten long lines and reduce code repetition

- Split long if-statement into separate lines
- Move in-line dictionary access to variables
- Change order of logic to separate repeated code
- Replace long lambda functions with local function

* Split up long if-statement

* Replace long lambda with local function

* Split up long list comprehension

* Split up long function call

* Create local function to simplify play_game() call

* Make function for process count display

* Move comments closer to left edge

And delete comment that repeats what the code says.

* Split long function declaration into lines

* Replace nested function with argument list

That is, replace something that doesn't work with something that does.

* Combine challenge accept and decline

When checking if a challenge is supported in a bot's configuration,
return a tuple of True/False for whether the challenge is accepted and
a string/None for the reason it was rejected.

Also, support methods should take the configuration as an argument
instead of parsing it in the calling body.

* Shorten lines with dict.get() default arguments

* Shorten lines with string bool expressions

* Replace comment with variable name

* Replace redundant expression with constant

* Split up long if-statement

* Simplify correspondence disconnect timing

* Delete trailing whitespace

* Simplify move overhead calculation

Also, 1e6 is easier for a human to parse than counting zeros in 1000000.

* Fix URL for test player

* Use JSON function for encoding

* Replace long string literals with objects

Should these need to be edited, the programmer will have help from
the IDE and better error messages when running.

* Restore full process count message

* Function returns string instead of logging

Keep logging in the main start() function to avoid collisions
with logging messages from other threads.

* Better logging of ending exception

* Better logging of PGN writing errors

* Revert "Function returns string instead of logging"

This reverts commit 9ffde398fd6f87a7b91a59c8328e5143805b8748.

* Use standard exception logger for game errors

* Fix punctuation in config error

* Delete unused exception variable

* Use logger.exception to log exception