Use separate method for online chess DB access (#518)
* Use separate method for online chess DB access

Since the api_get() method uses backup to retry communications, any
problems with communicating with the online chess databases can
result in retrying for up to a minute. This would waste time that bots
could use to pick their own moves.

This commit replaces the Lichess.api_get() method with another that
does not have backoff, so it does not retry requests.

* Remove raise_for_status

The call to .json() will raise an error if no data is returned.

* Remove unnecessary arguments from api_get()

Nothing used the raise_for_status argument for api_get(), so it was
removed.

The get_game_pgn() was the only one to use the params argument,
but that argument had no effect, so it was removed along with
the params argument for api_get().