Filter matchmaking based on relative ratings (#536)
* Filter matchmaking based on relative ratings

The user may specify a maximum rating difference when searching
for an opponent. Only those bots with a rating less than this difference
in the game to be played will be chosen.

* Fix tests to accommodate new user profile info

* Use 100 for default rating difference

* Periodically update rating range

Every five minutes, download the bot's user profile so that the rating
differences filter keeps creating a good range as the bot's rating
changes.