Delay challenging bot after challenge decline (#563)
* Delay challenging bot after challenge decline

If an opponent declines a challenge, do not challenge them again for one
hour. Increase the delay by another hour on every subsequent decline.

Will close #538.

* More fine-grained delay control

Whether a bot is challenged again depends not just on whether that bot
declines a challenge, but also on the type of game that was declined.
Lichess-bot will delay challenging a bot if that bot has previously refused
a game with the same chess variant, time control speed, and rated option.

* Unify logic for parameters where "random" is valid

* Do not add self to delay register

* Correct grammar of comment

* Separate delay filtering

Keep a list of bots prior to filtering by delay timers. If the delay filtering
results in no available bots to challenge, try challenging a bot from the
prior list.

* Set delay on usernames only

* Make post-decline delays optional

* Document delay option in README

* More README details for decline delays

* Revert "Set delay on usernames only"

This reverts commit f8f8bbdd9084b371de018e9d19bec0f09ef89d3c.

* User can choose type of delay

* Use proper code formatting in README