Run fewer GitHub actions (#639)
GitHub actions will now only be run when there is a PR, or the push commit message contains the text Update version (which it already does because we use Update version to x.y.z.n). This is done to avoid having to rerun as many failures. Also, update LC0 version, and use DEBUG instead of INFO.