Handle errors from Lichess class outside of games (#658)
* Ignore errors from Lichess.decline_challenge()

The challenge will expire or the challenger will cancel the challenge
eventually.

* Ignore errors when updating user profile

Since this is only used to update the engines rating when looking for
similarly rated opponents, missing one update is not critical since it
is repeated every 5 minutes.

The other call to Lichess.get_profile() in start_lichess_bot() is unchanged
since an error here stops the program before anything starts. This may be
an early warning that something is wrong with lichess.org. It is quick for the
user to restart the program.

* Add missing return to decline_challenge()

* Handle when getting PGN from lichess fails

In this case, a failure to download the PGN from lichess will return an
empty string. This will cause the PGN file to be written from local
information. The header of the PGN will be missing information
(perhaps this can be rectified in the future), but the moves and engine
output will be written. The file name will contain enough information
to manually download the game from lichess.