Fix difference in Linux/Windows handling of cwd (#710)
* Fix difference in Linux/Windows handling of cwd

When a user has the engine: working_dir: field set in their config
file, the behavior of popen when opening the engine executable
is different on Windows and Linux. In Linux, the current working
directory (cwd) is changed before starting the executable. In
Windows, it's the reverse. This means that Linux clients will not
be able to find the executable if engine: working_dir is different
than the cwd when lichess-bot starts.

These changes fix the problem by making the path to the engine
absolute before starting the process. Also, the documentation is
updated to notify users that the cwd change means relative paths
to files are now relative to the working_dir.

Closes #707

* Fix incorrect cwd warning

* Update config.yml.default