Allow for toggle-type engine options (#812)
An engine may have a command line option that doesn't take a
value--i.e., one that toggles some option on or off. This will be
indicated in the user's config file by leaving the value of an option
empty or nil. This change will translate the engine_options section
of their config file from this

    engine:
      name: engine_name
      engine_options:
        settings: settings.txt
        logging:

into the command line instruction

    engine_name --settings_file=settings.txt --logging