Only use local_game_done events to signal games are over (#847)
* Only respond to local_game_done event

Instead of making local changes to the lichess-bot state based on
receiving either gameFinish or local_game_done (removing game ID
from active games and signaling to matchmaker than a game has
ended), wait for a local_game_done event only. This guarantees that
the bot engine process has completed.

The other change required is to put a local_game_done event in the
control_queue when the game process terminates due to an
exception, i.e., before the game process can add the local_game_done
event after creating the PGN record.

* Attempt to get game PGN on in-game error

The engine search information is lost, but the moves, clock, and ending
information will be there.

Also, delete unneeded inner function doc and add typing information.

* Get game information from downloaded PGN

If the game process ended due to an error, then the Game state
won't be available. Extract what information there is from the PGN
and write to a single-game file (as if
config.pgn_file_grouping == "game") since the game is likely
unfinished.

* Get game information from downloaded PGN

If the game process ended due to an error, then the Game state
won't be available. Extract what information there is from the PGN
and write to a single-game file (as if
config.pgn_file_grouping == "game") since the game is likely
unfinished.

* Add more detail to save_pgn_record() docs

* Do not attempt to print PGN on errors

* Get game info and path from PGN if game error

* Simplify PGN transport

All information needed to write the PGN file is in the local_game_done
event. The information is either loaded at the end of a game in the
final_queue_entries() function, or, in case of an error during the game,
in the game_error_handler() inside start_game_thread(). The
model.Game instance from inside the game thead is no longer needed.

* Fix PGN player names in test game

* Delete impossible condition