Matchmaking pull back (#901)
* Don't blank matchmaker ID in start_game()

This is already done in Matchmaker.accepted_challenge()

* Use challenge.from_self to filter challenges

A bot does not need to respond to challenges from itself.
Instead of checking Matchmaker.challenge_id, use
Challenge.from_self. Plus, this change handles the case where
if the bot's user manually creates a challenge for the bot on
the lichess website.

* Collect repeated code into method

This method names the purpose of setting challenge_id
to the empty string.