Quit lichess-bot after all games finish (#902)
* Option to quit lichess-bot after all games complete

* Fix thread_logging_configurer typing

* Use correct argument for queue.get()

* Use pool.close() and join() to stop new games

* Fix missing log messages

It seems that problems occur when Ctrl-C is pressed while queue.get() is
blocking. Afterwards, every other log message is missing while calling
queue.task_done() results in exceptions being raised. This change calls
queue.get() without blocking to prevent these problems.

* Revert unnecessary changes

* Delete terminating log message

The program is shutting down, no need to announce it. Plus, with the
switching of log methods, this message is likely to be missed due to
the logging_listener_proc() process being terminated.

* Pressing Ctrl-C twice quits immediately

In case the user needs to quit lichess-bot immediately, they
can now press Ctrl-C twice to override the wait-until-games-finish option.

* Handle None from control_queue after SIGTERM

In Linux, when SIGTERM is sent to lichess-bot, control_queue.get()
returns None, which causes an exception to be raised for the line:
if "type" not in event:

* Update wiki regarding quitting immediately

* Ensure log messages from game threads are handled

In case final game thread exits while logging_listener_proc() is in a
timeout sleep.

* Prevent multiple SIGINT messages from logging

* Print message when waiting for games to end

* Update wiki regarding quitting

* Reduce complexity of lichess_bot_main()

* Prevent Ctrl-C from terminating engine