Allow the bot to prioritize playing against humans or bots (#979)
Allow the bot to prioritize playing against humans or bots. For example, if someone changes preference to human and has sort_by set to best, then the bot will prefer to play with a human even if they are rated lower than a bot. The order of opponents will be:

HUMAN A (2100)
HUMAN B (1800)
HUMAN C (1500)
BOT A (3000)
BOT B (2500)